requirements.txt
# EAN / product lookup
httpx>=0.27.0

# Image search: SerpAPI (primary), DuckDuckGo (fallback)
google-search-results>=2.4.2
duckduckgo-search>=4.0.0

# Claude Vision
anthropic>=0.39.0

# Env & config
python-dotenv>=1.0.0

# Obrazy: pomniejszanie przed zapisem do bazy
Pillow>=10.0.0

# Baza: Vercel Postgres / Neon (Postgres)
psycopg[binary]>=3.1.0
